Streamlined typescript files
