+++ resources/ts/modules/pksanc/box.ts
@@ -1,7 +1,5 @@
 import {DataCardlist} from "../../components/datalists/DataCardlist";
-import {DataSelect} from "../../components/datalists/DataSelect";
 import {openModal, init as initModals, closeModal} from "../../components/modal";
-import {IronbrainError} from "../../Exceptions/IronbrainError";
 
 /**
  * Initializes the pksanc box page
