+++ resources/ts/modules/pksanc/deposit.ts
@@ -1,7 +1,6 @@
 import {closeModal, init as initModals} from "../../components/modal";
 import {IronbrainError} from "../../Exceptions/IronbrainError";
-import {toast} from "../../main";
-import {postData} from "../../ajax";
+import {postData, toast} from "../../main";
 import {DataSelect} from "../../components/datalists/DataSelect";
 
 const romhackModalId: string = 'add_romhack_modal';
