--- resources/ts/utilities.ts
@@ -1,8 +1,0 @@
-export function toDisplayString(string:string):string {
-    string = string.replace('_', ' ');
-    return string;
-}
-
-export function toast(string:string) {
-    console.log(string)
-}
